# services/catalog/internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Catalog Service API
  description: Categories and products for the e-commerce marketplace
  version: 1.0.0
tags:
  - name: General
    description: Health and service status
  - name: Categories
    description: Category CRUD and tree navigation
  - name: Products
    description: Product CRUD and listing
servers:
  - url: http://localhost:8081
    description: Local (direct)
  - url: http://localhost:8000/v1/catalog
    description: Via Kong
paths:
  /health:
    get:
      tags: [General]
      summary: Health check
      operationId: health
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: catalog
  /v1/categories:
    post:
      tags: [Categories]
      summary: Create a category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name:
                  type: string
                  example: Eletrônicos
                slug:
                  type: string
                  example: eletronicos
                parent_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Categories]
      summary: List categories
      operationId: listCategories
      parameters:
        - name: parent_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by parent category (omit for root categories)
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /v1/categories/tree:
    get:
      tags: [Categories]
      summary: List categories as nested tree
      operationId: listCategoriesTree
      description: Returns the full category hierarchy with children nested. Use for menus and navigation.
      responses:
        "200":
          description: Nested category tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryTreeNode"
  /v1/categories/{id}:
    get:
      tags: [Categories]
      summary: Get category by ID
      operationId: getCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [Categories]
      summary: Update category
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name:
                  type: string
                slug:
                  type: string
                parent_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Validation error
        "404":
          description: Category not found
        "409":
          description: Slug already exists
    delete:
      tags: [Categories]
      summary: Delete category
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Category deleted
        "404":
          description: Category not found
  /v1/products:
    post:
      tags: [Products]
      summary: Create a product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
    get:
      tags: [Products]
      summary: List products with pagination and filters
      operationId: listProducts
      parameters:
        - name: seller_id
          in: query
          schema:
            type: string
            format: uuid
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
                  total:
                    type: integer
  /v1/products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    patch:
      tags: [Products]
      summary: Update product
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
      responses:
        "200":
          description: Product updated
        "400":
          description: Validation error
        "404":
          description: Product not found
    delete:
      tags: [Products]
      summary: Delete product
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found
components:
  schemas:
    CategoryTreeNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        parent_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        children:
          type: array
          items:
            $ref: "#/components/schemas/CategoryTreeNode"
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        parent_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        seller_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        price:
          type: string
        images:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProduct:
      type: object
      required: [seller_id, category_id, title, description, price]
      properties:
        seller_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        price:
          type: string
        images:
          type: array
          items:
            type: string
    Error:
      type: object
      properties:
        error:
          type: string